Grant users in the the sudo group remote login access
This gives the sudo group the same access as the admin group.

http://askubuntu.com/questions/43317/what-is-the-difference-between-the-sudo-and-admin-group
"Up until Ubuntu 11.10, administrator access using the sudo tool was
granted via the admin Unix group. In Ubuntu 12.04, administrator
access will be granted via the sudo group. This makes Ubuntu more
consistent with the upstream implementation and Debian."